+++ sqllogictest-engine/src/postgres.rs
@@ -1,30 +1,27 @@
-use std::sync::Arc;
+mod extended;
+mod simple;
 
-use async_trait::async_trait;
-use sqllogictest::{ColumnType, DBOutput};
+use std::{marker::PhantomData, sync::Arc};
+
 use tokio::task::JoinHandle;
-
-use crate::DBConfig;
 
 pub type Result<T> = std::result::Result<T, tokio_postgres::Error>;
 
-pub struct Postgres {
+pub struct Simple;
+pub struct Extended;
+
+pub struct Postgres<P> {
     client: Arc<tokio_postgres::Client>,
     join_handle: JoinHandle<()>,
+    protocol: PhantomData<P>,
 }
 
-impl Postgres {
-    pub async fn connect(config: &DBConfig) -> Result<Self> {
-        let (host, port) = config.random_addr();
+pub type PostgresSimple = Postgres<Simple>;
+pub type PostgresExtended = Postgres<Extended>;
 
-        let (client, connection) = tokio_postgres::Config::new()
-            .host(host)
-            .port(port)
-            .dbname(&config.db)
-            .user(&config.user)
-            .password(&config.pass)
-            .connect(tokio_postgres::NoTls)
-            .await?;
+impl<P> Postgres<P> {
+    pub async fn connect(config: tokio_postgres::Config) -> Result<Self> {
+        let (client, connection) = config.connect(tokio_postgres::NoTls).await?;
 
         let join_handle = tokio::spawn(async move {
             if let Err(e) = connection.await {
@@ -35,6 +32,7 @@ impl Postgres {
         Ok(Self {
             client: Arc::new(client),
             join_handle,
+            protocol: PhantomData,
         })
     }
 
@@ -44,64 +42,8 @@ impl Postgres {
     }
 }
 
-impl Drop for Postgres {
+impl<P> Drop for Postgres<P> {
     fn drop(&mut self) {
         self.join_handle.abort()
-    }
-}
-
-#[async_trait]
-impl sqllogictest::AsyncDB for Postgres {
-    type Error = tokio_postgres::error::Error;
-
-    async fn run(&mut self, sql: &str) -> Result<DBOutput> {
-        let mut output = vec![];
-
-        // NOTE:
-        // We use `simple_query` API which returns the query results as strings.
-        // This means that we can not reformat values based on their type,
-        // and we have to follow the format given by the specific database (pg).
-        // For example, postgres will output `t` as true and `f` as false,
-        // thus we have to write `t`/`f` in the expected results.
-        let rows = self.client.simple_query(sql).await?;
-        let mut cnt = 0;
-        for row in rows {
-            let mut row_vec = vec![];
-            match row {
-                tokio_postgres::SimpleQueryMessage::Row(row) => {
-                    for i in 0..row.len() {
-                        match row.get(i) {
-                            Some(v) => {
-                                if v.is_empty() {
-                                    row_vec.push("(empty)".to_string());
-                                } else {
-                                    row_vec.push(v.to_string());
-                                }
-                            }
-                            None => row_vec.push("NULL".to_string()),
-                        }
-                    }
-                }
-                tokio_postgres::SimpleQueryMessage::CommandComplete(cnt_) => {
-                    cnt = cnt_;
-                    break;
-                }
-                _ => unreachable!(),
-            }
-            output.push(row_vec);
-        }
-
-        if output.is_empty() {
-            Ok(DBOutput::StatementComplete(cnt))
-        } else {
-            Ok(DBOutput::Rows {
-                types: vec![ColumnType::Any; output[0].len()],
-                rows: output,
-            })
-        }
-    }
-
-    fn engine_name(&self) -> &str {
-        "postgres"
     }
 }
